Fix build-static.sh error
Thanks for the great work! This works fantastically in Kate.

I was trying to run the build-static.sh script, and it prints out
```
Building qml-lsp...
no Go files in <cloned project path>
Building qml-lint...
Building qml-refactor-fairy...
```
and the qml-lsp-static file was not created.
Fixing the path in build-static.sh solves this issue.